requirements.txt
boto3>=1.26.0
botocore>=1.29.0
pytest>=7.3.1
moto[dynamodb,sts]>=4.1.11
python-dotenv>=1.0.0
pydantic>=2.4.2
tenacity>=8.2.3
pytest-cov>=4.1.0